pyproject.toml
[project]
name = "guppi"
version = "0.1.0"
description = "A Python CLI agentic assistant built on the Anthropic API."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Rob Prouse" }]

dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pydantic-ai>=0.0.14",
    "rich>=13.7",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "mcp>=1.0",
]

[project.scripts]
guppi = "guppi.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guppi"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
